+++ apps/community-bot/src/server.ts
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright 2026 Sovri SAS
+
+import { run } from "probot";
+
+import registerCommunityBot from "./app.js";
+
+run(registerCommunityBot);